requirements.txt
langchain==0.3.25
langgraph==0.2.70
langchain-openai==0.3.18
openai>=1.50.0
requests==2.32.3
supabase==2.15.0
pydantic==2.11.4
tiktoken==0.9.0
python-dotenv==1.1.0
tenacity==9.1.2
